output adjustment linked in group

+++ fluorender/FluoRender/Fui/OutAdjustPanel.cpp
@@ -422,7 +422,7 @@ void OutAdjustPanel::OnRGammaText(wxCommandEvent& event)
 	double val;
 	str.ToDouble(&val);
 	m_r_gamma_sldr->SetValue(int(val * 100));
-	m_agent->setValue("gamma r", 1.0/val);
+	m_agent->setValue("gamma r", double(1.0/val));
 }
 
 void OutAdjustPanel::OnGGammaChange(wxScrollEvent & event)
@@ -438,7 +438,7 @@ void OutAdjustPanel::OnGGammaText(wxCommandEvent& event)
 	double val;
 	str.ToDouble(&val);
 	m_g_gamma_sldr->SetValue(int(val * 100));
-	m_agent->setValue("gamma g", 1.0/val);
+	m_agent->setValue("gamma g", double(1.0/val));
 }
 
 void OutAdjustPanel::OnBGammaChange(wxScrollEvent & event)
@@ -454,7 +454,7 @@ void OutAdjustPanel::OnBGammaText(wxCommandEvent& event)
 	double val;
 	str.ToDouble(&val);
 	m_b_gamma_sldr->SetValue(int(val * 100));
-	m_agent->setValue("gamma b", 1.0/val);
+	m_agent->setValue("gamma b", double(1.0/val));
 }
 
 //brightness
@@ -471,7 +471,7 @@ void OutAdjustPanel::OnRBrightnessText(wxCommandEvent& event)
 	double val;
 	str.ToDouble(&val);
 	m_r_brightness_sldr->SetValue(int(val));
-	m_agent->setValue("brightness r", val/256.0+1.0);
+	m_agent->setValue("brightness r", double(val/256.0+1.0));
 }
 
 void OutAdjustPanel::OnGBrightnessChange(wxScrollEvent & event)
@@ -487,7 +487,7 @@ void OutAdjustPanel::OnGBrightnessText(wxCommandEvent& event)
 	double val;
 	str.ToDouble(&val);
 	m_g_brightness_sldr->SetValue(int(val));
-	m_agent->setValue("brightness g", val/256.0+1.0);
+	m_agent->setValue("brightness g", double(val/256.0+1.0));
 }
 
 void OutAdjustPanel::OnBBrightnessChange(wxScrollEvent & event)
@@ -503,7 +503,7 @@ void OutAdjustPanel::OnBBrightnessText(wxCommandEvent& event)
 	double val;
 	str.ToDouble(&val);
 	m_b_brightness_sldr->SetValue(int(val));
-	m_agent->setValue("brightness b", val/256.0+1.0);
+	m_agent->setValue("brightness b", double(val/256.0+1.0));
 }
 
 void OutAdjustPanel::OnRHdrChange(wxScrollEvent &event)
